springboot-mvc: validation
Here I am creating an application for Customer to enter his/her firstName and lastName in a form. Here lastName is mandatory so it cannot be empty. If form get successfully validated, it would redirect the user to the confirmation screen where I am just displaying the entered values.

The main thing here is the validation. Spring framework provides us with so many validation annotations. here I am just using NotNull and Size validation.

Customer.Java: first step is to add the validation on the mandatory field which in this case is lastName using annotation.

customer-form.html: here I am creating a simple page with one form of user related fields. And all the databinding and other concepts I have already covered in 06-01-springboot-mvc-demo. Here tricky part is the error handling. "${#fields.hasErrors('lastName')}" here '#fields' have all the fields of Customer class. and the whole line is checking if the lastName field has an error in it. If the pervious condition is successful, then in next line "th:errors="*{lastName}" it is looking for the corresponding error which ever was triggered due to the incorrect input. And all this will automatically handle, thanks to springframework.

CustomerController.java: here in landingFormProcess, @Valid annotation is telling the code to perform the validation on the Corresponding object which in this case is customer. And the result of this validation is going to store in BindingResult object. later in the method, then I am simply checking if bindingRresult has some error or not. If there is an error, then simply redirect the user to the customer-form. or in case of successful validation, take the user to the confirmation screen.