springboot-mvc-validation: initbinder
There was an issue in my previous commit. if I entered the blank spaces in the lastName field. my validations were getting success for it. Which is not correct in practical examples. So to handle this scenario,  I am using initBinder. i have added all the necessary details about initBinder in the class file as comments.