springboot-mvc-validation: addtional validation annotation
Here I am just adding few more validation annotations to build a better vocabulary for using them in real world examples. I am adding another file here that is messages.properties. I ll explain it at the end of this commit details.

Customer.Java: @Min as the name states, check the lower bound of the value. @Max checks the upper bound of the value. @Pattern annotation use to perform regex validation. And in this example, this regex restrict the input to 5 chars which can either be digits or chars (both lower & upper case) but can not be special characters.

customer-confirmation.html: I am just showing that additional data here on this page that I added in Customer Object.

messages.properties: Now this is interesting. If you just open the customer-form.html page, here you ll see that I am using the input type 'number'. The reason behind is obvious that the field in which it is going to store the data is integer too. So this field won't even allow the user to enter any other character except digits. Now what if, for some reason I don't want to restrict the user to enter only digits. in that case when user enters the other characters in the field, System will show the long exception infront of the field. So to handle this error properly and show the user appropriate error message, I looked into the bindingResult object in CustomerController and print it in the logs. There is would be showing something like this "typeMismatch.customer.freePasses,typeMismatch.freePasses". Spring iss showing the error corresponding to this key "typeMismatch.customer.freePasses,". So I am simply overriding the the message for this key in message.properties file. but to work this solution properly, you need to make sure the exact file name along with the exact error key are being used